fix PluginLifecycleListener classes.
