+++ PluginsAndFeatures/azure-toolkit-for-intellij/azure-intellij-plugin-hdinsight-base/src/main/java/com/microsoft/intellij/AzureActionsListener.java
@@ -12,6 +12,7 @@ import com.microsoft.azure.cosmosspark.serverexplore.cosmossparknode.CosmosSpark
 import com.microsoft.azure.hdinsight.common.HDInsightHelperImpl;
 import com.microsoft.azure.hdinsight.common.HDInsightLoader;
 import com.microsoft.azure.toolkit.intellij.common.CommonConst;
+import com.microsoft.azure.toolkit.lib.Azure;
 import com.microsoft.azure.toolkit.lib.common.messager.ExceptionNotification;
 import com.microsoft.azure.toolkit.lib.common.operation.AzureOperation;
 import com.microsoft.azuretools.authmanage.CommonSettings;
@@ -69,6 +70,7 @@ public class AzureActionsListener implements AppLifecycleListener, PluginCompone
                 new CosmosSparkClusterOpsCtrl(CosmosSparkClusterOps.getInstance()));
 
             ServiceManager.setServiceProvider(TrustStrategy.class, IdeaTrustStrategy.INSTANCE);
+            CommonSettings.setUserAgent(Azure.az().config().getUserAgent());
             initAuthManage();
         } catch (final Throwable t) {
             log.error(t.getMessage(), t);
